Update StyleManager.swift
The original ButtonStyles in SwiftUI are limited to a few classic looks and behaviours. However, in my project, I need more sophisticated style controlling mechanisms, such as hovering and disabling controls. So, I created a set of customized button style values. It can deliver a set of easy control styles to light up buttons on both iOS and macOS.